fix(type_reachability): enum/typedef-aware record_index; split leaf helpers
Codex review, two more findings on the record-provenance mechanism
(6885dd7):

1. _spelling_index's enum-collision guard only covered a record's derived
   suffix. A record whose own full identity carries no namespace at all
   (already bare, structurally indistinguishable from a derived suffix)
   was never checked against the same enum-collision vocabulary --
   mine::Wrapper became just "Wrapper" as a global record, and a bare
   global record "Wrapper" colliding with an unrelated enum "other::
   Wrapper" resolved unambiguously to the record regardless.

2. The record index was built entirely independently of the typedef
   vocabulary, so a record's exact/global identity could unconditionally
   shadow a typedef alias reducing to the identical spelling --
   _typedef_spelling_targets already drops its own candidate for this
   collision (checked one-directionally, against records only), but the
   record side won by default regardless.

Both fixed by checking a record's own exact-identity registration (not
just its derived suffixes, which were already checked) against the same
enum-aware and new typedef-aware (_raw_typedef_spellings, a fresh,
unfiltered candidate vocabulary distinct from the already-filtered
_typedef_candidate_spellings) collision sets. Confirmed empirically with
reproductions matching both reported scenarios.

Also splits type_reachability.py's pure spelling/matching helpers (no
scan state) into a new type_reachability_spelling.py leaf module -- this
round of hardening pushed the file past the AI-readiness 2000-line hard
cap. One-directional dependency (type_reachability.py imports from the
new module, never the reverse), every name re-exported for backward
compatibility with dumper_scoping.py/export_surface.py/existing tests
that import private helpers directly from abicheck.type_reachability.
Purely mechanical, no behavior change -- confirmed via mypy (explicit
re-export aliases satisfy --no-implicit-reexport) and the full targeted
test suite.

verify.py --profile fast, mypy, check_ai_readiness.py (0 errors, file-size
cap resolved), check_fp_rate.py, and measure_contract_shadow.py all pass.
Regression tests added for both new findings.
