+++ changelog.d/20260809_record_enum_typedef_collision_fixes.md
@@ -1,0 +1,4 @@
+### Fixed
+
+- **Closed five more ambiguity gaps in `type_reachability.directly_referenced_stdlib_type_spellings`'s record-reachability provenance** (Codex review, multiple rounds): a record reached via an ambiguous typedef alias, and only *later* upgraded to a trustworthy route discovered *during the record walk itself* (not during seeding), previously kept its stale, ambiguous-only provenance forever due to the worklist's "queue once" dedup — `_StdlibReferenceScan.reach_record` now requeues an already-walked record when its provenance is genuinely upgraded. `_spelling_index`'s own `record_index` construction is now enum-aware (a record whose bare spelling collides with an unrelated enum's own spelling is no longer treated as an unconditionally trustworthy direct match, including when the record's own full identity — not just a derived suffix — is the one that's already bare) and typedef-aware (a record's spelling colliding with a typedef key or its own derived suffix no longer wins by default over the typedef, mirroring how `_typedef_spelling_targets` already drops its own candidate for the identical collision). An already-resolved typedef alias's cached provenance now also propagates to any record its own chain terminates at, not just to stdlib identities found directly in the chain text. All confirmed empirically with reproductions matching each reported scenario, including order-independence.
+- **Split `type_reachability.py`'s pure spelling/matching helpers into a new `type_reachability_spelling.py` leaf module** to stay under the AI-readiness file-size hard cap after this round of hardening. Purely mechanical (all names re-exported from `type_reachability.py` for backward compatibility) — no behavior change.
